# Rust quorum intersection checker integration