--- a/src/herder/HerderUtils.cpp
+++ b/src/herder/HerderUtils.cpp
@@ -3,10 +3,12 @@
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
 #include "herder/HerderUtils.h"
+#include "crypto/KeyUtils.h"
+#include "main/Config.h"
+#include "rust/RustVecXdrMarshal.h"
 #include "scp/Slot.h"
 #include "xdr/Stellar-ledger.h"
 #include <algorithm>
-#include <xdrpp/marshal.h>
 
 namespace stellar
 {
@@ -40,4 +42,50 @@ getStellarValues(SCPStatement const& statement)
 
     return result;
 }
+
+// Render `id` as a short, human readable string. If `cfg` has a value, this
+// function uses `cfg` to render the string. Otherwise, it returns the first 5
+// hex values `id`.
+std::string
+toShortString(std::optional<Config> const& cfg, NodeID const& id)
+{
+    if (cfg)
+    {
+        return cfg->toShortString(id);
+    }
+    else
+    {
+        return KeyUtils::toShortString(id).substr(0, 5);
+    }
+}
+
+QuorumIntersectionChecker::QuorumSetMap
+toQuorumIntersectionMap(QuorumTracker::QuorumMap const& qmap)
+{
+    QuorumIntersectionChecker::QuorumSetMap ret;
+    for (auto const& elem : qmap)
+    {
+        ret[elem.first] = elem.second.mQuorumSet;
+    }
+    return ret;
+}
+
+std::pair<std::vector<PublicKey>, std::vector<PublicKey>>
+toQuorumSplitNodeIDs(QuorumSplit& split)
+{
+    std::vector<NodeID> leftNodes;
+    leftNodes.reserve(split.left.size());
+    for (const auto& str : split.left)
+    {
+        leftNodes.push_back(KeyUtils::fromStrKey<NodeID>(std::string(str)));
+    }
+    std::vector<NodeID> rightNodes;
+    rightNodes.reserve(split.right.size());
+    for (const auto& str : split.right)
+    {
+        rightNodes.push_back(KeyUtils::fromStrKey<NodeID>(std::string(str)));
+    }
+    return std::make_pair(std::move(leftNodes), std::move(rightNodes));
+}
+
 }