--- a/src/main/Config.h
+++ b/src/main/Config.h
@@ -677,6 +677,9 @@ class Config : public std::enable_shared_from_this<Config>
     // Whether to run online quorum intersection checks.
     bool QUORUM_INTERSECTION_CHECKER;
 
+    // Whether to use the new Rust SAT-solving based quorum intersection
+    // checker.
+    bool USE_QUORUM_INTERSECTION_CHECKER_V2;
     // Invariants
     std::vector<std::string> INVARIANT_CHECKS;