… Cargo.lock
… docs/versioning-soroban.md
… src/herder/test/QuorumIntersectionTests.cpp
… src/rust/src/quorum_checker.rs