# Description

This is the implementation of a new quorum intersection checker that leverages a SAT-solver. It contains

- The bridge implementation for `stellar/stellar-quorum-analyzer`, the underneath Rust library.  
- A new adaptor class `RustQuorumCheckerAdaptor` that wraps the Rust calls in C++ that provides some basic metrics. 
- Async interruption handling
- A new config flag `USE_QUORUM_INTERSECTION_CHECKER_V2` to toggle between old and new solver.
- Update all existing test cases to pass with the new solver. 

Rust side: https://github.com/stellar/stellar-quorum-analyzer/pull/1

Preliminary benchmark results look very good. This is run with the stellar-core `check-quorum-intersection` command line locally, sample test cases can be found [here](https://github.com/jayz22/stellar-quorum-analyzer-fork/tree/quorum-checker/tests/test_data/random). A 16-org configuration which previously won't finish, now takes only fraction of a second. 

| no. of orgs / nodes  | old solve time(s) | new solve time(s) |
| ------------------- | ----------------- | ------------------ |
| 10 (30)                    | 1.59                      | 0.05                        |
| 12  (36)                   | 40.09                   | 0.01                       |
| 13  (39)                   | 5166.47               | 0.03                       |
| 14  (42)                   | 6469.11               | 0.03                       |
| 16  (48)                   | n/a                       | 0.12                        | 


Marking draft for now since it needs to be merged after the Rust side PR (and update the commit hash in cargo.toml). However, this it is now ready for review. 

# Checklist
- [x] Reviewed the [contributing](https://github.com/stellar/stellar-core/blob/master/CONTRIBUTING.md#submitting-changes) document
- [x] Rebased on top of master (no merge commits)
- [x] Ran `clang-format` v8.0.0 (via `make format` or the Visual Studio extension)
- [x] Compiles
- [x] Ran all tests
- [x] If change impacts performance, include supporting evidence per the [performance document](https://github.com/stellar/stellar-core/blob/master/performance-eval/performance-eval.md)