@jayz22 2025-02-24
Superseded by https://github.com/stellar/stellar-core/pull/4653, which provides process-isolation for extra safety.